Juuuuust getting started (otter, path, folder); Many todos; many ideas
All linked. Trying to be concise and yet have fun and indicative to myself what I was doing so I can be interested enough to pick it up tomorrow. Programming is the most boring exciting painful thing I could ever imagine doing and is only fun when you're in "the zone". Oh, and the zone is really hard to get into and any distractions kill it. Oh, and these thoughts don't stop - not for the people that make interesting stuff at least.

Oh by the way (THANKS PRIME A GEN) - I've got to stop typing now on my regular MacBook keyboard because my weak measly non-man hands hurt after 10 minutes of typing.